newUpdate
base bg
base btn
frm setup
vuex update
