+++ launcher/src/components/UI/BaseButton.vue
@@ -1,0 +1,26 @@
+<template>
+  <button id="BaseButton">
+    <slot></slot>
+  </button>
+</template>
+
+<script>
+export default {
+  name: "BaseButton",
+};
+</script>
+
+<style scoped>
+button {
+  border-radius: 40px;
+  background-color: rgb(43, 42, 40);
+  text-align: center;
+  font-weight: bold;
+  padding: 10px;
+  font-size: 16pt;
+  color: #eee;
+}
+button:hover {
+  background-color: grey;
+}
+</style>
