+++ launcher/src/main.js
@@ -1,24 +1,32 @@
-import Vue from "vue";
+import Vue from 'vue';
+import BaseButton from './components/UI/BaseButton.vue';
+import BackgroundPage from './components/layers/BackgroundPage.vue';
 import App from "./App.vue";
 import router from "./router";
 import store from "./store";
-import Toasted from "vue-toasted";
+// import Toasted from "vue-toasted";
 import ControlService from '@/store/ControlService';
-import { BootstrapVue, BootstrapVueIcons } from "bootstrap-vue";
-import "bootstrap/dist/css/bootstrap.css";
-import "bootstrap-vue/dist/bootstrap-vue.css";
+// import { BootstrapVue, BootstrapVueIcons } from "bootstrap-vue";
+// import "bootstrap/dist/css/bootstrap.css";
+// import "bootstrap-vue/dist/bootstrap-vue.css";
+// Vue.use(Toasted);
+// Vue.use(BootstrapVue);
+// Vue.use(BootstrapVueIcons);
+// import "./main.css";
+Vue.component('base-button',BaseButton);
+Vue.component('background-page',BackgroundPage);
+ new Vue({
+router,
+store,
+ render: (h) => h(App),
 
-Vue.use(Toasted);
-Vue.use(BootstrapVue);
-Vue.use(BootstrapVueIcons);
-import "./main.css";
+ }).$mount("#app");
 
-ControlService.init(store);
+//Vue.config.productionTip = false;
 
-Vue.config.productionTip = false;
+// new Vue({
+//   router,
+//   store,
+//   render: (h) => h(App),
+// }).$mount("#app");
 
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount("#app");
