+++ launcher/src/pages/BaseHome.vue
@@ -6,18 +6,19 @@
 <script>
 import TheFirst from "./TheFirst.vue";
 import SetupServer from "./SetupServer.vue";
+import WelcomePage from "./WelcomePage.vue";
+
 export default {
   data() {
     return {
       pageName: "TheFirst",
     };
   },
-  components: { TheFirst, SetupServer },
+  components: { TheFirst, SetupServer, WelcomePage },
   name: "BaseHome",
   methods: {
     page(pageId) {
       this.pageName = pageId;
-      alert(this.pageName);
     },
   },
 };
