+++ launcher/src/pages/WelcomePage.vue
@@ -1,0 +1,13 @@
+<template>
+  <section>
+    <background-page>Welcome</background-page>
+  </section>
+</template>
+<script>
+import BackgroundPage from "../components/layers/BackgroundPage.vue";
+export default {
+  components: { BackgroundPage },
+  name: "WelcomePage",
+};
+</script>
+<style scoped></style>
