+++ launcher/src/store/index.js
@@ -1,11 +1,16 @@
-import Vue from "vue";
-import Vuex from "vuex";
+import Vue from 'vue'
+import Vuex from 'vuex'
 
-Vue.use(Vuex);
+Vue.use(Vuex)
 
-export default new Vuex.Store({
-  state: {},
-  mutations: {},
-  actions: {},
-  modules: {},
-});
+const store = new Vuex.Store({
+  
+  state: {
+    
+  },
+  mutations: {
+    increment (state) {
+      state.count++
+    }
+  }
+})
